Refactors AddOnPin component to isolate pinning behavior
