+++ src/addons/browser/AddOnListItem.svelte
@@ -1,83 +1,18 @@
-<script lang="ts" context="module">
-  import { writable, type Writable } from "svelte/store";
-
-  interface Author {
-    name?: string;
-    avatar?: string;
-  }
-
-  // API endpoint https://api.www.documentcloud.org/api/addons/
-  export interface AddOnListItem {
-    id: number;
-    name: string;
-    repository: string;
-    parameters: any;
-    description?: string;
-    author?: Author;
-    usage?: number;
-    categories: string[];
-    documents: string[];
-    active: boolean;
-    featured: boolean;
-    default: boolean;
-  }
-
-  export const pinned: Writable<AddOnListItem[]> = writable([]);
-</script>
-
 <script lang="ts">
   import { _ } from "svelte-i18n";
-  import { baseApiUrl } from "../../api/base.js";
-  import { getCsrfToken } from "../../api/session.js";
 
-  import Pin from "../../common/Pin.svelte";
+  import AddOnPin from "../AddOnPin.svelte";
   import AddOnPopularity from "../Popularity.svelte";
+  import type { AddOnListItem } from "../types.js";
 
   export let addon: AddOnListItem;
 
-  $: endpoint = new URL(`/api/addons/${addon.id}/`, baseApiUrl);
   $: description = addon.parameters?.description;
   $: if (!addon?.author) {
     addon.author = { name: addon?.repository?.split("/")[0] };
   }
 
   $: url = `#add-ons/${addon.repository}`;
-
-  async function toggle(event) {
-    event.preventDefault();
-
-    const csrftoken = getCsrfToken();
-    const options: RequestInit = {
-      credentials: "include",
-      method: "PATCH", // this component can only update whether an addon is active or not
-      headers: { "X-CSRFToken": csrftoken, "Content-type": "application/json" },
-    };
-
-    // optimistic update
-    addon.active = !addon.active;
-
-    const resp = await fetch(endpoint, {
-      ...options,
-      body: JSON.stringify({ active: addon.active }),
-    }).catch((err) => {
-      addon.active = !addon.active;
-      return {
-        ok: false,
-        statusText: String(err),
-      };
-    });
-
-    if (!resp.ok) {
-      // reset active state
-      addon.active = !addon.active;
-      console.error(`Problem updating add-on: ${resp.statusText}`);
-    }
-
-    // now that we've updated, set $pinned
-    $pinned = addon.active
-      ? [...$pinned, addon]
-      : $pinned.filter((a) => a.id !== addon.id);
-  }
 </script>
 
 <style>
@@ -146,7 +81,7 @@
   <div class="container" id={addon.repository}>
     <div class="top-row">
       <div class="center-self">
-        <Pin active={addon.active} on:click={toggle} />
+        <AddOnPin {addon} />
       </div>
       <div class="stretch">
         <h3 class="addon-name">{addon.name}</h3>
